Fix cp signature checking.
When we receive packages to a package store, we can only
accept signed packages. Before we were just verifying the
package stream via a challenge, however this is insufficient,
as once the challenge was verified an attacker may start
to send unsigned tarballs.

The new protocol essentially works like this:

-> Send signed package closure.
<- Reply with filtered package closure.
-> Send signed tarballs, one per package.